refactor: Todo API 예외 처리

+++ src/todo/todo.service.ts
@@ -1,5 +1,5 @@
 import { TodoDto } from './dto/todo.dto';
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { Todo } from './todo.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
@@ -8,11 +8,12 @@ import { Repository } from 'typeorm';
 export class TodoService {
   constructor(
     @InjectRepository(Todo)
-    private readonly todoRepository: Repository<Todo>,
+    private readonly todoRepository: Repository<Todo>
   ) {}
   // create todo
   async createTodo(todoDto: TodoDto): Promise<void> {
     await this.todoRepository.create(todoDto).save();
+    return;
   }
   // get todo list
   async getTodoList(): Promise<Todo[]> {
@@ -20,17 +21,31 @@ export class TodoService {
   }
   // delete todo
   async deleteTodo(id: number): Promise<void> {
+    const data = await this.todoRepository.findOneBy({ id });
+    if (!data) {
+      throw new BadRequestException();
+    }
     await this.todoRepository.delete({ id });
+    return;
   }
   // update todo
   async updateTodo(id: number, todoDto: TodoDto): Promise<void> {
     const { content } = todoDto;
+    const data = await this.todoRepository.findOneBy({ id });
+    if (!data) {
+      throw new BadRequestException();
+    }
     await this.todoRepository.update(id, { content });
+    return;
   }
   // update todo status
   async updateStatus(id: number): Promise<boolean> {
-    const ex = await this.todoRepository.findOneBy({ id });
-    await this.todoRepository.update(id, { isCompleted: !ex.isCompleted });
-    return !ex.isCompleted;
+    const data = await this.todoRepository.findOneBy({ id });
+    if (!data) {
+      throw new BadRequestException();
+    }
+    console.log(!data.isCompleted);
+    await this.todoRepository.update(id, { isCompleted: !data.isCompleted });
+    return !data.isCompleted;
   }
 }
